Implement a mechanism to be able to commit messages
The ce sdk is commiting the messages once the receiver handler exits,
but because we pass the ce into a channel we need to find a way make the
handler wait until we finished process the message.
This commit add a small entity with the ce and a channel inside. The
channel is closed once we finished the processing.

Signed-off-by: Cosmin Tupangiu <cosmin@redhat.com>